Added a few features, lots of docs
* Changed run behavior -- endpoints and paths can now be specified
within the file. This makes cli commands less unwieldy. This isn't
all that expected of a CLI tool, but I think the purpose of the script
is to make things as easy as possible to understand so people can get
up and running as quickly as possible. If they knew the Globus ins and
outs, they wouldn't need this script.
* Added Client Credentials authentication.
* Added lots of docs